Extract decorator into standaline file
