+++ packages/studio-base/src/components/AppBar/User.stories.tsx
@@ -2,7 +2,7 @@
 // License, v2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
-import { StorybookDecorator } from "./StorybookDecorator";
+import { StorybookDecorator } from "./StorybookDecorator.stories";
 import { UserMenu } from "./User";
 
 export default {
